Move infoPopUps to popUps folder and fix paths in index
