+++ features/facebookImport/src/popUps/baseInfoPopUp/baseInfoPopUp.css
@@ -1,10 +1,12 @@
 .base-info {
+    position: absolute;
     background-color: var(--background-light);
     color: var(--text-dark);
     height: 100%;
     margin: 0 auto;
     overflow: hidden scroll;
     padding: 16px;
+    z-index: 100;
 }
 
 .base-info .info-title {
