+++ features/facebookImport/src/popUps/index.js
@@ -1,10 +1,10 @@
-import DataStructureInfoScreen from "./views/infoScreens/dataStructureInfoScreen/dataStructureInfoScreen.jsx";
-import ActivitiesInfoScreen from "./views/infoScreens/activitiesInfoScreen/activitiesInfoScreen.jsx";
-import MessagesInfoScreen from "./views/infoScreens/messagesInfoScreen/messagesInfoScreen.jsx";
-import PicturesInfoScreen from "./views/infoScreens/picturesInfoScreen/picturesInfoScreen.jsx";
-import PostReactionInfoScreen from "./views/infoScreens/postReactionInfoScreen/postReactionInfoScreen.jsx";
-import OnOffFacebookInfoScreen from "./views/infoScreens/onOffFacebookInfoScreen/onOffFacebookInfoScreen.jsx";
-import OffFacebookInfoScreen from "./views/infoScreens/onOffFacebookInfoScreen/offFacebookInfoScreen.jsx";
+import DataStructureInfoScreen from "./infoPopUps/dataStructureInfo.jsx";
+import ActivitiesInfoScreen from "./infoPopUps/activitiesInfo.jsx";
+import MessagesInfoScreen from "./infoPopUps/messagesInfo.jsx";
+import PicturesInfoScreen from "./infoPopUps/picturesInfo.jsx";
+import PostReactionInfoScreen from "./infoPopUps/postReactionInfo.jsx";
+import OnOffFacebookInfoScreen from "./infoPopUps/onOffFacebookInfo.jsx";
+import OffFacebookInfoScreen from "./infoPopUps/offFacebookInfo.jsx";
 
 export default {
     "data-structure-info": DataStructureInfoScreen,
