+++ features/facebookImport/src/popUps/infoPopUps/dataStructureInfo.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 
-import i18n from "../../../i18n.js";
-import BaseInfoPopUp from "../../../components/baseInfoPopUp/baseInfoPopUp.jsx";
-import Infographic from "../../../components/infographic/infographic.jsx";
+import i18n from "../../i18n.js";
+import BaseInfoPopUp from "../baseInfoPopUp/baseInfoPopUp.jsx";
+import Infographic from "../../components/infographic/infographic.jsx";
 
-import "../infoPopUps.css";
+import "./infoPopUps.css";
 
 const DataStructureInfoScreen = () => {
     const dataStructureInfoText = [
