+++ features/facebookImport/src/popUps/infoPopUps/offFacebookInfo.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-import i18n from "../../../i18n.js";
-import BaseInfoPopUp from "../../../components/baseInfoPopUp/baseInfoPopUp.jsx";
-import Infographic from "../../../components/infographic/infographic.jsx";
+import i18n from "../../i18n.js";
+import BaseInfoPopUp from "../baseInfoPopUp/baseInfoPopUp.jsx";
+import Infographic from "../../components/infographic/infographic.jsx";
 
 const OffFacebookInfoScreen = () => {
     const offFacebookInfoText = [
