+++ features/facebookImport/src/popUps/infoPopUps/picturesInfo.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import i18n from "../../../i18n.js";
-import BaseInfoPopUp from "../../../components/baseInfoPopUp/baseInfoPopUp.jsx";
+import i18n from "../../i18n.js";
+import BaseInfoPopUp from "../baseInfoPopUp/baseInfoPopUp.jsx";
 
 const PicturesInfoScreen = () => {
     const picturesInfoText = [
