+++ features/facebookImport/src/popUps/infoPopUps/postReactionInfo.jsx
@@ -1,11 +1,12 @@
 import React from "react";
 
-import i18n from "../../../i18n.js";
-import BaseInfoPopUp from "../../../components/baseInfoPopUp/baseInfoPopUp.jsx";
-import Infographic from "../../../components/infographic/infographic.jsx";
+import i18n from "../../i18n.js";
+import BaseInfoPopUp from "../baseInfoPopUp/baseInfoPopUp.jsx";
+import Infographic from "../../components/infographic/infographic.jsx";
 
-import "./postReactionInfoScreen.css";
-import "../infoScreens.css";
+import "./postReactionInfo.css";
+import "./infoPopUps.css";
+
 const PostReactionInfoScreen = () => {
     const postReactionInfoText = [
         <>
